+++ pochidetection/datasets/transforms/__init__.py
@@ -1,0 +1,5 @@
+"""学習時 transforms (letterbox 等)."""
+
+from pochidetection.datasets.transforms.letterbox import LetterboxTransform
+
+__all__ = ["LetterboxTransform"]